+++ npm/designsystem/src/__mocks__/useHover.ts
@@ -1,0 +1,7 @@
+jest.mock('../hooks/useHover', () => ({
+  useHover: jest.fn().mockImplementation(ref => {
+    return { hoverRef: ref || { current: undefined }, isHovered: false };
+  }),
+}));
+
+export {};