+++ npm/designsystem/src/__mocks__/useOutsideEvent.ts
@@ -1,0 +1,5 @@
+jest.mock('../hooks/useOutsideEvent', () => ({
+  useOutsideEvent: jest.fn(),
+}));
+
+export {};