+++ npm/designsystem/src/utils/tests/setup-test.ts
@@ -4,3 +4,5 @@ import '../../__mocks__/matchMedia';
 import '../../__mocks__/IntersectionObserver';
 import '../../__mocks__/ResizeObserver';
 import '../../__mocks__/useLayoutEvent';
+import '../../__mocks__/useSize';
+import '../../__mocks__/useHover';